Remove old mainTag when switching preset, even when disused:
Currently if you switch the preset of a disused amentity, the old mainTag will still remain.

For example if a node has `disused:amenity=cafe`, and you changed the preset to "Convenience Shop"
(without first pressing "Mark active"), you would end up with `disused:amenity=cafe` and
`shop=convenience` on the same node.

Therefore we should remove the `disused:` prefix if it exists, before we remove the old preset tags.